Improved test_installation check, compliant with the latest TorchX updates
