+++ README.MD
@@ -76,8 +76,9 @@ We are currently iterating on the setup experience. For now, we provide manual i
 
 5. Test the installation.
    ```
-   torchx run --scheduler local_cwd test_installation.py:test_installation
+   torchx run -s local_cwd --script test_installation.py
    ```
+   See [TorchX](https://pytorch.org/torchx/) for more information on launching distributed and remote jobs.
 
 6. If you want to run a more complex example, please take a look at the torchrec [DLRM example](examples/dlrm/dlrm_main.py).
